refactor: use daisyui steps for audiobook pairing
